+++ lib/pro_image_editor.dart
@@ -59,6 +59,7 @@ export 'shared/widgets/reactive_widgets/reactive_custom_appbar.dart';
 export 'shared/widgets/reactive_widgets/reactive_custom_widget.dart';
 export 'shared/widgets/extended/extended_pop_scope.dart';
 export 'core/constants/editor_style_constants.dart';
+export 'core/utils/image_converter.dart';
 
 /// Widgets
 export 'shared/widgets/animated/fade_in_up.dart';